feat(playground): add page-level tabs for tables
